requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.11
gunicorn==20.1.0
pycodestyle==2.8.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
django-crispy-forms==1.13.0
django-simple-menu==1.2.1
django-tables2
django-filter